Finally done with the basics
Lexer code generation exists, but it's completely wrong and should never be used.
But the lexer generation itself should work nicely apart from it not getting optimized just yet, that's the next plan of action.
I just realized there's no real reason for me to write all this, but here it is xD